+++ src/process-commands/rss.ts
@@ -19,8 +19,10 @@ import { downloadAudio } from '../process-steps/02-download-audio'
 import { runTranscription } from '../process-steps/03-run-transcription'
 import { selectPrompts } from '../process-steps/04-select-prompt'
 import { runLLM } from '../process-steps/05-run-llm'
-import { filterRSSItems, saveAudio, saveInfo, parser } from '../utils/validate-option'
-import { l, err, logSeparator, logInitialFunctionCall, logRSSProcessingStatus } from '../utils/logging'
+import { saveAudio, saveInfo, parser } from '../utils/validate-option'
+import { filterRSSItems } from '../utils/rss-utils'
+import { l, err, logSeparator, logInitialFunctionCall } from '../utils/logging'
+import { logRSSProcessingStatus } from '../utils/rss-utils'
 import { retryRSSFetch } from '../utils/retry'
 
 import type { ProcessingOptions, RSSItem } from '../utils/types/process'
